Update Capacitor and Qonversion dependencies to latest versions
This commit updates the Capacitor and Qonversion dependencies in the package.json and Capacitor build.gradle files. The Capacitor version has been updated from 6.0.0 to 7.2.0, and the QonversionCapacitorPlugin version has also been updated from 0.1.3 to 0.2.1. Additionally, the Java version in the android/app/capacitor.build.gradle file has been increased from 17 to 21, and the CocoaPods version has been updated from 1.15.2 to 1.16.2. These updates ensure compatibility with the latest versions of Capacitor and Qonversion plugins, enhancing the stability and functionality of the project.